refactor: make ErrorMessage, View
