+++ src/main/kotlin/racingcar/util/ErrorMessage.kt
@@ -1,0 +1,8 @@
+package racingcar.util
+
+object ErrorMessage {
+    const val RUNCATCHING_NO_RESULT = "runCatching didn't success and didn't fail also"
+    const val FAIL_TO_READ_INPUT: String = "입력에 오류가 발생했습니다. 다시 입력해주세요."
+    const val INVALID_ROUND: String = "시도할 횟수는 1 이상의 자연수여야 합니다. 다시 입력해주세요."
+    const val OVER_MAX_CAR_NAME_LENGTH: String = "자동차 이름은 5자를 초과할 수 없습니다. 다시 입력해주세요."
+}
