+++ src/main/kotlin/racingcar/util/extension/StringExtension.kt
@@ -1,3 +1,3 @@
-package racingcar.extension
+package racingcar.util.extension
 
 operator fun String.times(times: Int): String = List(times) { this }.joinToString(separator = "")
